requirements.txt
joblib=1.4.2
matplotlib=3.9.4
networkx=3.2.1
numpy=1.23.2
pandas=2.2.3
python=3.9.21
torch=2.1.0
torch_geometric=2.5.2
pytorch-scatter=2.1.2
pytorch-sparse=0.6.18
scikit-learn=1.6.1
scipy=1.13.1
tqdm=4.67.1

